PAGOPA-848 refactoring old to new
