+++ src/main/java/it/gov/pagopa/payments/controller/receipt/impl/PaymentsController.java
@@ -2,12 +2,10 @@ package it.gov.pagopa.payments.controller.receipt.impl;
 
 import it.gov.pagopa.payments.controller.receipt.IPaymentsController;
 import it.gov.pagopa.payments.entity.ReceiptEntity;
-import it.gov.pagopa.payments.entity.ReceiptEntityCosmos;
 import it.gov.pagopa.payments.model.PaymentsResult;
 import it.gov.pagopa.payments.model.ReceiptModelResponse;
 import it.gov.pagopa.payments.model.ReceiptsInfo;
 import it.gov.pagopa.payments.service.PaymentsService;
-import it.gov.pagopa.payments.service.PaymentsServiceCosmos;
 import it.gov.pagopa.payments.utils.CommonUtil;
 import java.util.stream.Collectors;
 import javax.validation.Valid;
@@ -33,8 +31,7 @@ public class PaymentsController implements IPaymentsController {
 
   @Autowired private ModelMapper modelMapper;
 
-  @Autowired private PaymentsService paymentsService;
-  @Autowired private PaymentsServiceCosmos paymentsServiceCosmos;
+  @Autowired private PaymentsService paymentsServiceCosmos;
 
   @Override
   public ResponseEntity<String> getReceiptByIUV(String organizationFiscalCode, String iuv) {
@@ -44,7 +41,7 @@ public class PaymentsController implements IPaymentsController {
             "GET",
             "getReceiptByIUV",
             String.format(LOG_BASE_PARAMS_DETAIL, organizationFiscalCode) + "; iuv= " + iuv));
-    ReceiptEntityCosmos receipt =
+    ReceiptEntity receipt =
         paymentsServiceCosmos.getReceiptByOrganizationFCAndIUV(organizationFiscalCode, iuv);
     return new ResponseEntity<>(receipt.getDocument(), HttpStatus.OK);
   }
@@ -68,7 +65,7 @@ public class PaymentsController implements IPaymentsController {
                 + debtor
                 + "; service= "
                 + service));
-    PaymentsResult<ReceiptEntityCosmos> receipts =
+    PaymentsResult<ReceiptEntity> receipts =
         paymentsServiceCosmos.getOrganizationReceipts(organizationFiscalCode, debtor, service, from, to);
     return new ResponseEntity<>(
         ReceiptsInfo.builder()
