+++ src/test/java/it/gov/pagopa/payments/modelmapper/ConvertReceiptEntityToReceiptModelResponseTest.java
@@ -24,6 +24,6 @@ class ConvertReceiptEntityToReceiptModelResponseTest {
 
     ReceiptModelResponse res = modelMapper.map(re, ReceiptModelResponse.class);
     assertNotNull(res);
-    assertEquals(re.getRowKey(), res.getIuv());
+    assertEquals(re.getOrganizationFiscalCode(), res.getIuv());
   }
 }
